Implement lexical subroutines.
This is done by automatically remove mysubs from the namespace at the end of
the scope they've been declared in by replacing the CV's GV with a copy without
the CV value. If a function with the name of the mysub existed before, the
original one will be restored after cleaning the mysub.