Make the peep recurse via PL_peepp
Also allows extensions, when delegating to Perl_peep, to specify what function
it should use when recursing into a part of the op tree.

The usecase for this are extensions like namespace::alias, which need to hook
into the peep to do their thing. With this change they can stop copying the
whole peep only to add tiny bits of new behaviour to it, allowing them to work
easier on a large variety of perls, without having to maintain one peep which
works on all of them (which is HARD!).